requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pyjwt
openpyxl
sqlalchemy
pandas
python-dotenv
qrcode
python-jose[cryptography]
reportlab
